Remove double SelectTypeahead component
