+++ src/utils/components/SelectTypeahead/utils/utils.ts
@@ -3,7 +3,7 @@ import { isEmpty } from '@kubevirt-utils/utils/utils';
 import { SelectOptionProps } from '@patternfly/react-core';
 
 import { CREATE_NEW } from './constants';
-export const createItemId = (value: any) => `select-typeahead-${value.replace(' ', '-')}`;
+export const createItemId = (value: any) => `select-typeahead-${value?.replace(' ', '-')}`;
 
 export const getCreateNewFolderOption = (
   filterValue: string,
